Allow get_remote_user to fetch username from Authorization header.
The new behavior can be configured with the setting `own_home.remote_user
= 'authorization'`. With the multiplication of remote_user getter,
`own_home.remote_user = (session|header|authorization)` should be used instead of
`own_home.get_username_from_session.enabled`. The current version adds
some logic to stay backward compatible.

This patch is usefull when using Own Home with Searchguard and the
`Authorization` header can be trusted. This method disable the use of
groups, but remove issues where the Own Home session is not in sync with
SearchGuard. It also allow the login to redirect anywhere instead of the
Own Home application.